requirements.txt
PyFxA==0.0.1
pytest-mozwebqa
unittestzero
